% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/au_theme.R
\name{au_minimal_theme}
\alias{au_minimal_theme}
\title{AU's ggplot theme}
\usage{
au_minimal_theme()
}
\value{
a ggplot theme for the AU
}
\description{
AU's ggplot theme
}
\examples{
library(ggplot2)


table(sample(LETTERS[1:6],
               size = 100,
               replace = TRUE)) \%>\%
    as.data.frame() \%>\%
  # create a bar graph with the fake data
  ggplot(aes(x = Var1,
             y = Freq,
             fill = Var1)) +
  geom_bar(stat = "identity",
           position = position_dodge(.8)) +
  scale_y_continuous("Count") +
  scale_x_discrete("") +
  au_minimal_theme()
}
