% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary_to_numeric.R
\name{binary_to_numeric}
\alias{binary_to_numeric}
\title{Re-code binary response into numeric values}
\usage{
binary_to_numeric(response)
}
\arguments{
\item{response}{name of the vector or column containing character strings of "yes" and "no"}
}
\value{
a numeric value of 0 for "no" and 1 for "yes"
}
\description{
Re-code binary response into numeric values
}
\examples{
response = c(sample(c("yes", "no"), 6, replace = TRUE), "YES")
binary_to_numeric(response)
}
