% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ca_doc}
\alias{ca_doc}
\title{Documentation of Common Assessment (CA) Tools}
\format{
A data frame containing metadata of CA data and information about CAT
\describe{
\item{QUESTION_SHORT_EN}{Computer-friendly name of CA questions}
\item{QUESTION_SHORT_ALIAS_EN}{Human-friendly CA questions}
\item{MODULE_SEQUENCE}{}
\item{GROUP_SEQUENCE}{}
\item{QUESTION_SEQUENCE}{}
\item{EST_CA_TEMPLATE_ID}{}
\item{VERSION_NUMBER}{}
\item{EFFECTIVE_FROM_DATE}{}
\item{MODULE_NAME_EN}{}
\item{MODULE_NAME_FR}{}
\item{GROUP_DESC}{}
\item{GROUP_DESC_FR}{}
\item{EST_CA_QUESTION_ID}{}
\item{PARENT_QUESTION_CODE}{}
\item{PARENT_QUESTION_ID}{}
\item{QUESTION_CODE_ID}{}
\item{UPDATED_DATE}{}
\item{CREATED_DATE}{}
\item{EST_CA_DATA_TYPE_NAME}{}
\item{QUESTION_EN}{}
\item{QUESTION_FR}{}
\item{QUESTION_DESC_EN}{}
\item{QUESTION_DESC_FR}{}
\item{TOOLTIP_EN}{}
\item{TOOLTIP_FR}{}
}
}
\usage{
ca_doc
}
\description{
Documentation of CA data containing metadata and CAT questions.
Metadata was created using \code{data-raw/CA_Current_Questions_and_Definitions.sql}.
}
\examples{
  ca_doc
}
\keyword{datasets}
