% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_query.R
\name{clean_query}
\alias{clean_query}
\title{Clean SQL query}
\usage{
clean_query(text_or_path = "//example/path/to/some_query.sql")
}
\arguments{
\item{text_or_path}{a text SQL query or path to a sql file}
}
\value{
A single string output
}
\description{
A function developed by Tori Oblad; See the original \href{https://stackoverflow.com/a/58446028}{thread} for details.
}
\examples{


query <- "
/* this query has
intentionally messy
comments
*/

Select
COL_A -- with a comment here
,COL_B
,COL_C
FROM
-- and some helpful comment here
Database.Datatable
;
-- or wherever

/* and some more comments here */

"

clean_query(query)

}
\references{
\url{https://stackoverflow.com/a/58446028}
}
\author{
Tori Oblad
}
