% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_validate.R
\name{cross_validate}
\alias{cross_validate}
\title{Cross validate categorical coding of transformed variables}
\usage{
cross_validate(
  data,
  codebook = FALSE,
  suffix = c("yn", "num", "stdz", "freq", "recency")
)
}
\arguments{
\item{data}{dataset of original and transformed variables to be cross validated}

\item{codebook}{if TRUE, the function produces a codebook of
original and transformed variables; if false, the function returns data frames of each transformed variable}

\item{suffix}{a string vector of suffix(es) used to name transformed variables.}
}
\value{
data frame(s) or a codebook
}
\description{
Cross validate categorical coding of transformed variables
}
