% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{est_catchment}
\alias{est_catchment}
\title{EST Catchment Areas by Phase}
\format{
\subsection{\code{est_catchment}}{

A table with 16 rows and 4 columns:
\describe{
\item{catchment}{EST catchment area name}
\item{phase}{Phase of transformation}
\item{SSM}{SSM in corresponding catchment area}
}
}
}
\source{
Analytics Unit > Projects > EST > Catchment Areas
}
\usage{
est_catchment
}
\description{
Reference table of EST catchment areas and their phase and SSMs
}
\keyword{datasets}
