% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{geo_ref}
\alias{geo_ref}
\title{Geographical reference table}
\format{
\subsection{\code{geo_ref}}{

A table with 16 rows and 4 columns:
\describe{
\item{catchment_id}{Catchment ID}
\item{catchment_area}{EST catchment area name}
\item{geo_uid}{StatCan's geographical region ID}
\item{economic_region}{Economic region}
}
}
}
\source{
Analytics Unit > Projects > EST > Catchment Areas > 2021 Provincial Rollout > Geo_Reference.xlsx
}
\usage{
geo_ref
}
\description{
Reference table for mapping \code{geo_uid} from StatCan data to EST catchment areas and economic regions
}
\keyword{datasets}
