% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data.R
\name{import_data}
\alias{import_data}
\title{Import data into R session}
\usage{
import_data(file_path = NULL)
}
\arguments{
\item{file_path}{Absolute path of the file. It will be prompted to choose file interactively if missing}
}
\value{
File as R object
}
\description{
Allows you to import a file either using an user-specified file path or interactively.
}
\examples{
\dontrun{
data = import_data()
}

}
