% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_ca_connection.R
\name{make_ca_connection}
\alias{make_ca_connection}
\title{Make connection to CDR}
\usage{
make_ca_connection(dsn = "CDR")
}
\arguments{
\item{dsn}{name of the DB's service name (SERVICE_NAME) to connect; the default is CDR}
}
\value{
a connection object to Centralized Data Repository
}
\description{
Creates a connection object using an ODBC driver with the log-in credentials saved in the system environment variables as "userid" (user ID) and "pwd" (password); check this instruction:
\code{vignette("Use ODBC login as environment variables", package = "autemplates")}
}
\examples{
make_ca_connection(dsn = "CDR")

}
