% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{noc_2021}
\alias{noc_2021}
\title{NOC 2021 Table}
\format{
A data frame with x rows and y column(s)
\describe{
\item{NOC_CODE}{A 5-digit NOC code associated with an occupation}
\item{NOC}{Job title}
\item{NOC_1_DIGIT}{1st digit of NOC code}
\item{BROAD_CATEGORY}{The Broad Category (first digit) of the classification represents the occupational categorization which is defined by the type of work performed, the field of study, or the industry of employment. There are 10 Broad categories in NOC 2021 Version 1.0.}
\item{NOC_2_DIGIT}{First two digits of NOC code}
\item{MAJOR_GROUP}{The Major Group (first and second digits) of the classification is represented by the Broad occupational categorization (first digit) and TEER categorization (second digit) together. A major group also encompasses several sub-major groups and thus represents the two-digit code used by the NOC. There are 45 major groups in NOC 2021 Version 1.0.}
\item{NOC_3_DIGIT}{First three digits of NOC code}
\item{SUB_MAJOR_GROUP}{The Sub-major Group (3-digit) of the classification represents the aggregation of several minor groups and thus represents the three-digit code used by the NOC. There are 89 sub-major groups in NOC 2021 Version 1.0.}
\item{NOC_4_DIGIT}{First four digits of NOC code}
\item{MINOR_GROUP}{The Minor Group (4-digit) of the classification represents the domain in which an occupation is carried out (occupational domain). It is an aggregation of several unit groups and thus represents the four-digit code used by the NOC. There are 162 minor groups in NOC 2021 Version 1.0.}
\item{NOC_TEER}{Second-digit of all codes}
\item{TEER_CATEGORY}{The TEER Category (second digit) of the classification represents the necessary training, education, experience and responsibilities of the occupation. There are 6 TEER categories in NOC 2021 Version 1.0.}
\item{TEER_DESCRIPTION}{The degree of Training, Education, Experience and Responsibilities.}
}
}
\source{
\url{https://github.com/ETDAU/trf-analysis/blob/8ea37620771743fa7b05561e4dd268bd8cec6e34/NOC_NAICS_Processing_Srcript.Rmd/"}
}
\usage{
noc_2021
}
\description{
Translation table for 2021 NOC codes
}
\keyword{datasets}
