% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/num_pct.R
\name{num_pct}
\alias{num_pct}
\title{Quickly compute the absolute numbers and proportions of unique values}
\usage{
num_pct(data, x)
}
\arguments{
\item{data}{name of the data frame}

\item{x}{name of the variable}
}
\value{
A table with the absolute number (n) and proportion (pct) of each level
}
\description{
Quickly compute the absolute numbers and proportions of unique values
}
\examples{

data.frame(letters = sample(LETTERS[1:6], size = 100, replace = T)) \%>\%
num_pct(., letters)

}
