% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rm_dup_responses.R
\name{rm_dup_responses}
\alias{rm_dup_responses}
\title{Remove duplicated entries in a response}
\usage{
rm_dup_responses(response, sep = ",")
}
\arguments{
\item{response}{name of the column that contains duplicated responses}

\item{sep}{separator of the duplicated entries; default is comma (",")}
}
\value{
a single response without a duplicate
}
\description{
Removes repeated responses that are concatenated together (e.g., "yes, yes")
by retaining only the first occurrence of the response
}
\note{
deprecated; only evaluates one word at a time
}
\examples{
duplicated_response = c("yes, YES", "no", "yes, no", "maybe true, Maybe true")
rm_dup_responses(response = duplicated_response, sep = ",")

}
